backend/requirements.txt
fastapi[all]
sqlalchemy
alembic
passlib
bcrypt
python-jose
python-dotenv